requirements.txt
# 核心依赖
streamlit>=1.28.0
langchain>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# 模型相关
modelscope>=1.9.0
transformers>=4.40.0
accelerate>=0.20.0
optimum>=1.16.0
auto-gptq>=0.7.0  # 4-bit量化支持

# 数据处理
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0

# 其他
torch>=2.0.0